Make some fields pub, reorg
